backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "verity-backend"
version = "0.1.0"
description = "VERITY — Autonomous Equity Research Agent Platform"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Async HTTP
    "httpx>=0.27.0",
    "aiohttp>=3.9.5",

    # LLM / Agent orchestration
    "langchain>=0.2.0",
    "langchain-anthropic>=0.1.15",
    "langchain-openai>=0.1.8",
    "langgraph>=0.1.0",
    "anthropic>=0.28.0",
    "openai>=1.30.0",

    # Vector DB
    "pinecone-client>=3.2.2",

    # Data / validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # Database
    "supabase>=2.4.6",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "alembic>=1.13.1",

    # Task queue / cache
    "celery[redis]>=5.4.0",
    "redis>=5.0.4",
    "upstash-redis>=1.1.0",

    # Financial data
    "yfinance>=0.2.40",
    "pandas>=2.2.2",
    "numpy>=1.26.4",

    # Document processing
    "tiktoken>=0.7.0",
    "pypdf>=4.2.0",
    "python-docx>=1.1.2",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.2",

    # Logging / observability
    "structlog>=24.2.0",
    "python-json-logger>=2.0.7",

    # Auth / security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "slowapi>=0.1.9",

    # Utilities
    "tenacity>=8.3.0",
    "python-dotenv>=1.0.1",
    "arrow>=1.3.0",
    "rich>=13.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",  # for TestClient
    "ruff>=0.4.7",
    "mypy>=1.10.0",
    "pre-commit>=3.7.1",
    "factory-boy>=3.3.0",
    "respx>=0.21.1",  # mock httpx
]

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "C4", "SIM"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.ruff.per-file-ignores]
"tests/**" = ["ANN"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=xml"

[tool.hatch.build.targets.wheel]
packages = ["app"]
